feat: refactor and run prettier on all files

+++ app/api/users/[id]/route.ts
@@ -2,10 +2,13 @@ import { prisma } from '@/lib/db'
 import { NextRequest, NextResponse } from 'next/server'
 
 /** Récupère un utilisateur via son identifiant */
-export async function GET({ params }: { params: Promise<{ id: string }> }) {
-    const id = (await params).id
-    const user = await prisma.user.findUnique({ where: { id } })
+export async function GET(request: NextRequest) {
+    const id = request.nextUrl.searchParams.get('id')
+    if (!id) {
+        return NextResponse.json({ error: 'Paramter "id" is required' }, { status: 400 })
+    }
 
+    const user = await prisma.user.findUnique({ where: { id } })
     if (!user) {
         return NextResponse.json({ error: 'User not found' }, { status: 404 })
     }
@@ -14,8 +17,12 @@ export async function GET({ params }: { params: Promise<{ id: string }> }) {
 }
 
 /** Modifie un utilisateur via son identifiant,*/
-export async function PATCH(request: NextRequest, { params }: { params: Promise<{ id: string }> }) {
-    const id = (await params).id
+export async function PATCH(request: NextRequest) {
+    const id = request.nextUrl.searchParams.get('id')
+    if (!id) {
+        return NextResponse.json({ error: 'Parameter "id" is required' }, { status: 400 })
+    }
+
     try {
         const body = await request.json()
         const user = await prisma.user.update({
@@ -32,8 +39,12 @@ export async function PATCH(request: NextRequest, { params }: { params: Promise<
 }
 
 /** Supprime un utilisateur via son identifiant */
-export async function DELETE({ params }: { params: Promise<{ id: string }> }) {
-    const id = (await params).id
+export async function DELETE(request: NextRequest) {
+    const id = request.nextUrl.searchParams.get('id')
+    if (!id) {
+        return NextResponse.json({ error: 'Parameter "id" is required' }, { status: 400 })
+    }
+
     try {
         await prisma.user.delete({ where: { id } })
